typecheck: fix assertion patterns checking
Before this commit assertion patterns are only checked to have the
required type. That is not enough, as the value provided in the
assertion pattern is used to determine the result type of the clause, so
we have to ensure the value provided in the assertion pattern is indeed
the only option, otherwise the type checking will become unsound
(see src/Dependent/EqAny.sfp for an example).

This patch adds checking for the assertion patterns. To achieve this we
return new meta variable as the pattern value and save a pair (meta,
assertion pattern value) for the later. By the end of clause pattern
checking this new meta variable should be solved to exactly the value
provided, otherwise it's an error.

This fixes all the affected variants EXCEPT Dependent.Monadic, since we
rely on the unification to solve the constraint for us.